+++ testsuite/tests/generics/T10604/T10604_no_PolyKinds.hs
@@ -1,0 +1,7 @@
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE KindSignatures #-}
+module T10604_no_PolyKinds where
+
+import GHC.Generics
+
+data F (f :: * -> *) = F deriving Generic1